pyproject.toml
[project]
name = "hyrax"
version = "0.1.0"
description = "Hyrax — a chaotic Telegram bot powered by Ollama"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "apscheduler>=3.10",
    "pydantic-settings>=2.4",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "pytest-mock>=3.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hyrax"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "pytest-mock>=3.14",
]
